Rename from "kubectl-split" to "kubectl-slice".
